requirements.txt
Django>=4.2
djangorestframework
django-environ
Pillow
pytest
pytest-django
drf-spectacular
